Create MovieGenre
This is a wee C# practice file I wanted to keep that suggests films once you have put in movie genres of your choice. It's a little buggy, sure, but I had to draw on a lot of new operations purely from forums on C# online and I'm proud I got it to work! Have at it if you'd like to improve it.